qa_assistant/requirements.txt
fastapi
uvicorn
slack_sdk
openai
langchain
langgraph
python-dotenv
jira
aiohttp
